feat: refactor location entity components
