+++ sites-config/features.json
@@ -103,34 +103,19 @@
         "dm_directoryParents.slug",
         "dm_baseEntityCount",
         "slug",
-        "c_bannerSection",
-        "c_heroSection",
-        "c_promoSection",
-        "c_featuredProductsSection.title",
-        "c_featuredProductsSection.products.name",
-        "c_featuredProductsSection.products.richTextDescription",
-        "c_featuredProductsSection.products.primaryPhoto",
-        "c_featuredProductsSection.products.c_primaryCTA",
-        "c_gallerySection",
         "c_aboutSection",
-        "c_teamSection.title",
-        "c_teamSection.team.id",
-        "c_teamSection.team.name",
-        "c_teamSection.team.headshot",
-        "c_teamSection.team.mainPhone",
-        "c_teamSection.team.c_occupation",
-        "c_teamSection.team.emails",
-        "c_teamSection.team.websiteUrl",
-        "c_faqSection.title",
-        "c_faqSection.faqs.question",
-        "c_faqSection.faqs.answer",
-        "c_nearbySection",
+        "c_bannerSection",
         "c_eventsSection.title",
         "c_eventsSection.events.name",
         "c_eventsSection.events.time",
         "c_eventsSection.events.description",
         "c_eventsSection.events.c_primaryCTA",
         "c_eventsSection.events.photoGallery",
+        "c_faqSection.title",
+        "c_faqSection.faqs.question",
+        "c_faqSection.faqs.answer",
+        "c_gallerySection",
+        "c_heroSection",
         "c_insightsSection.title",
         "c_insightsSection.cta",
         "c_insightsSection.insights.title",
@@ -140,8 +125,23 @@
         "c_insightsSection.insights.descriptionLong",
         "c_insightsSection.insights.descriptionShort",
         "c_insightsSection.insights.cta",
+        "c_nearbySection",
+        "c_featuredProductsSection.title",
+        "c_featuredProductsSection.products.name",
+        "c_featuredProductsSection.products.richTextDescription",
+        "c_featuredProductsSection.products.primaryPhoto",
+        "c_featuredProductsSection.products.c_primaryCTA",
+        "c_promoSection",
         "c_reviewsSection",
-        "ref_reviewsAgg.reviewCount"
+        "ref_reviewsAgg.reviewCount",
+        "c_teamSection.title",
+        "c_teamSection.team.id",
+        "c_teamSection.team.name",
+        "c_teamSection.team.headshot",
+        "c_teamSection.team.mainPhone",
+        "c_teamSection.team.c_occupation",
+        "c_teamSection.team.emails",
+        "c_teamSection.team.websiteUrl"
       ],
       "filter": {
         "entityTypes": ["location"]
