+++ src/components/cards/EventCard.tsx
@@ -2,7 +2,7 @@ import { Image, Link } from "@yext/pages/components";
 import { EventProfile } from "src/types/entities";
 import { CardComponent } from "src/models/cardComponent";
 
-export const EventCard: CardComponent<EventProfile> = function EventCard(
+const EventCard: CardComponent<EventProfile> = function EventCard(
   props
 ): JSX.Element {
   const { profile } = props;
@@ -57,3 +57,5 @@ function formatEventDate(startTime: string, endTime: string) {
     </div>
   );
 }
+
+export default EventCard;
