--- src/components/entity/Insights.css
@@ -1,11 +1,0 @@
-.Insights {
-  &-insightCard {
-    &:not(:first-child) {
-      padding-top: 32px;
-    }
-
-    &:not(:last-child) {
-      border-bottom: 1px solid theme(colors.brand-gray.300);
-    }
-  }
-}
