+++ src/styles/theme.css
@@ -1,11 +1,17 @@
+html {
+  --yxt-hoursTable-col2-width: auto;
+}
+
 .HoursStatus-current {
   font-weight: theme("fontWeight.bold");
 }
 
-.HoursTable-row {
-  margin-bottom: 4px;
+.HoursTable {
+  &-row {
+    margin-bottom: 4px;
+  }
 
-  &.is-today {
+  &-row.is-today {
     font-weight: theme(fontWeight.bold);
   }
 }
